dep: use initial rosu-pp rewrite (#638)
* dep: use initial rosu-pp rewrite

* fix: ran sqlx prepare

* chore: made clippy happy

+++ bathbot/src/active/impls/osustats/scores.rs
@@ -146,7 +146,7 @@ impl OsuStatsScoresPagination {
                     rank,
                     max_pp,
                     stars: attrs.stars() as f32,
-                    max_combo: attrs.max_combo() as u32,
+                    max_combo: attrs.max_combo(),
                 };
 
                 self.entries.insert(i, entry);
