+++ bathbot/src/commands/osu/pinned.rs
@@ -421,7 +421,7 @@ async fn process_scores(
         let mut calc = ctx.pp(&map).mode(score.mode).mods(&score.mods);
         let attrs = calc.difficulty().await;
         let stars = attrs.stars() as f32;
-        let max_combo = attrs.max_combo() as u32;
+        let max_combo = attrs.max_combo();
 
         let max_pp = match score
             .pp
